Added new functionalities to the validations package in goLF module
